data/requirements.txt
absl-py==1.3.0
astunparse==1.6.3
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
Cython==0.29.32
dill==0.3.6
dm-tree==0.1.7
efficientnet==1.1.1
etils==0.9.0
filelock==3.8.2
flatbuffers==1.12
fonttools==4.38.0
future @ file:///croot/future_1677599870788/work
gast==0.4.0
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
h5py==3.7.0
httplib2==0.21.0
huggingface-hub==0.11.1
idna @ file:///croot/idna_1666125576474/work
image-classifiers==1.0.0
imageio==2.28.0
immutabledict==2.2.3
importlib-metadata==5.1.0
importlib-resources==5.10.1
joblib==1.2.0
kaggle==1.5.12
keras==2.9.0
Keras-Applications==1.0.8
keras-efficientnet-v2==1.2.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
networkx==3.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.5.2.52
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
portalocker==2.6.0
promise==2.3
protobuf==3.19.6
psutil==5.9.4
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1670514731622/work
regex==2022.10.31
requests @ file:///croot/requests_1682607517574/work
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.3.1
scikit-image==0.20.0
scikit-learn==1.2.0
scipy==1.9.1
sentencepiece==0.1.97
seqeval==1.2.2
six==1.16.0
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-addons==0.19.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-metadata==1.12.0
tensorflow-model-optimization==0.7.3
tensorflow-text==2.9.0
termcolor==2.1.1
text-unidecode==1.3
tf-models-official==2.9.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.2
toml==0.10.2
torch==1.12.1
torchvision @ file:///croot/torchvision_1670313549871/work
tqdm==4.64.1
transformers==4.25.1
typeguard==2.13.3
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1680254681959/work
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
